requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.42
alembic==1.16.4
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic[email]==2.11.7
pydantic-settings==2.3.0 